Add terraform config for API proxy
We need to have an API proxy for TransportAPI so that we API credentials
are not exposed. In theory an API proxy also allows to have things like
rate limiting, which are not set up for now. The terraform config sets
up an AWS API Gateway that just passes all incoming requests to
TransportAPI, but adds credentials as query params.

The credentials (for both AWS and Transport API) are configured in a
terraform.tfvars file which is loaded by terraform automatically. There
is a template for that: terraform.tfvars.example.