Use API gateway in the code
Pass the URL of the API proxy from terraform as an environment variable
to webpack when it builds the bundle. Webpack will make the URL
available to the code as a constant.

When we run tests we don't use webpack, so we'll have to define the `API_URL` constant in Jest's config.