Force API gateway deployment on changes to infrastructure config
A new AWS API deployment is necessary every time the associated API
changes as otherwise the previous version will remain active at the
output api_url.

There is no standard way to do it via terraform and there is an open
issue about that. As a workaround we can rely the hash of the terraform
config file and force a new deployment every time the file changes. We
obviously don't need to re-deploy on every change, but given that the
configuration is minimal this easy workaround should be OK.